requirements.txt
Flask==3.1.2
pandas==2.3.2
python-docx==1.2.0
docxtpl==0.20.1
openpyxl==3.1.5
reportlab==4.4.4
docxcompose==1.4.0
python-dateutil==2.9.0.post0
psycopg2-binary>=2.9.9,<3.0
gunicorn==21.2.0
python-dotenv==1.0.0
